+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `array_get_arr` function
 - PHPUnit 9
 - Add `starts_with`, `ends_with` and `contains`
+- Add Klass module with `unqualified_name` function
 
 # 1.7.2
 - Experimental support for gRPC servers